backend/server/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
flask-login
pyjwt
bcrypt
python-dotenv
stripe
paypalrestsdk
pytest
pytest-cov
flask-testing
faker
requests
python-dateutil
email-validator
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1